fixed leak of JotView with the JotSDK registering and unregistering views #473

+++ Project/LooseLeaf/MMMemoryProfileView.m
@@ -50,10 +50,10 @@
     [[UIColor blackColor] setFill];
     
     // Drawing code
-    NSInteger numberInImageCache = [[MMLoadImageCache sharedInstace] numberOfItemsHeldInCache];
-    NSInteger numberOfLoadedPagePreviews = [[MMPageCacheManager sharedInstace] numberOfPagesWithLoadedPreviewImage];
-    NSInteger numberOfLoadedPageStates = [[MMPageCacheManager sharedInstace] numberOfStateLoadedPages];
-    NSInteger numberOfItemsInTrash = [[JotTrashManager sharedInstace] numberOfItemsInTrash];
+    int numberInImageCache = [[MMLoadImageCache sharedInstace] numberOfItemsHeldInCache];
+    int numberOfLoadedPagePreviews = [[MMPageCacheManager sharedInstace] numberOfPagesWithLoadedPreviewImage];
+    int numberOfLoadedPageStates = [[MMPageCacheManager sharedInstace] numberOfStateLoadedPages];
+    int numberOfItemsInTrash = [[JotTrashManager sharedInstace] numberOfItemsInTrash];
     
     CGFloat y = 50;
     [@"MMLoadImageCache:" drawAtPoint:CGPointMake(150, y) withFont:font];
